Try to fix dataSourceChanged signal not being properly connected
Probably was an issue with variable scope in the for loop connecting
the signals to layers. The partial() call might have stored a
reference to layer instead of the current value.  Now the code uses
layerid, which is a string value (immutable) returned from the layer
object instead of the object itself.
Not quite sure this was the reason, but now signals are properly
connected.